refactor(reco-core): extract the wgpu-free matrix leaf into src/geometry
Step 7 part 1: view_matrix, the clip constants, opengl_to_wgpu_matrix,
and matrix4_to_columns move out of the wgpu-heavy renderer into the new
geometry module (L0) - pure nalgebra, no logic change. The CPU stitch
geometry and the lens preview/undistort passes now import the shared
matrices from the leaf instead of reaching into the renderer;
build_gpu_uniforms stays render-side (it builds the shader interface).
