+++ crates/reco-core/src/lens/rig_correction.rs
@@ -233,7 +233,7 @@ mod tests {
     /// the wrong way).
     #[test]
     fn world_to_render_inverts_view_matrix_under_tilt_and_roll() {
-        use crate::render::renderer::view_matrix;
+        use crate::geometry::view_matrix;
         let position = [1.0_f32, 0.0, 1.0];
         let cam = VirtualCamera::new(&position);
         for &(tilt, roll) in &[
